// frontend/src/lib/chat-enums.ts
export enum chatEnum {
  error = "error-message",
  joinRoom = "join-room",
  joined='joined-user',
  disconnect = "disconnect",
  sendMessage = "sendMessage",
  receive = "receive_message",
  joinmeet = "join-meet",
  userConnected = "userConnected",
  signal = "signal",
  videoState='changevidostate',
  audioState='audiostatechange'

}